Add new Python application for timesheet download
This commit includes the creation of a new Python script that performs automated timesheet downloading. The script uses Playwright, a Python library, to interact with the Toggl website for fetching the timesheet data. In addition, several configuration and definition files related to the IntelliJ IDEA Python project setup are included.